Добавлен вспомогательный модуль для утилит
