+++ src/view/card-film.js
@@ -1,12 +1,12 @@
-import dayjs from 'dayjs';
+import {getCardDate} from '../utils.js';
 
 const MAX_LENGTH_DESCRIPTION = 140;
 
 const createCardFilmTemplate = (film) => {
-  const {name, rating, release, runtime, genres, img, description, comments, isWatchList, isWatched, isFavorite} = film;
+  const {name, rating, release, duration, genres, img, description, comments, isWatchList, isWatched, isFavorite} = film;
 
   const descriptionSliced = (description.length > 140) ? `${description.slice(0, MAX_LENGTH_DESCRIPTION - 1)}...` : description;
-  const date = dayjs(release).format('YYYY');
+  const date = getCardDate(release);
 
   return (
     `<article class="film-card">
@@ -14,7 +14,7 @@ const createCardFilmTemplate = (film) => {
       <p class="film-card__rating">${rating}</p>
       <p class="film-card__info">
         <span class="film-card__year">${date}</span>
-        <span class="film-card__duration">${runtime}</span>
+        <span class="film-card__duration">${duration}</span>
         <span class="film-card__genre">${genres}</span>
       </p>
       <img src="./images/posters/${img}" alt="" class="film-card__poster">
